Revert "1. add content-md5 header for  policy, request payment and versioning api. 2. split versionig test cases"
This reverts commit a66661d58c3a48f71fef4353a1c9ab404bc3094c.
