--- tests/.gitattributes
@@ -1,4 +1,0 @@
-invalid_sample_data.csv binary
-sample_data.csv binary
-sample_json.json binary
-sample_json_lines.json binary
