+++ tests/test_multipart.py
@@ -206,6 +206,137 @@ class TestMultipart(OssTestCase):
         self.assertEqual('中文', result.tag_set.tagging_rule[' + '])
 
         result = self.bucket.delete_object_tagging(key)
-   
+
+    def test_multipart_with_versionging(self):
+
+        from oss2.models import BucketVersioningConfig
+        from oss2.models import BatchDeleteObjectVersion
+        from oss2.models import BatchDeleteObjectVersionList
+
+        auth = oss2.Auth(OSS_ID, OSS_SECRET)
+        bucket_name = OSS_BUCKET + "-test-multipart-with-versionging"
+        bucket = oss2.Bucket(auth, OSS_ENDPOINT, bucket_name)
+
+        bucket.create_bucket(oss2.BUCKET_ACL_PRIVATE)
+
+        wait_meta_sync()
+
+        config = BucketVersioningConfig()
+        config.status = 'Enabled'
+
+        result = bucket.put_bucket_versioning(config)
+
+        wait_meta_sync()
+
+        result = bucket.get_bucket_info()
+
+        self.assertEqual(int(result.status)/100, 2)
+        self.assertEqual(result.bucket_encryption_rule.sse_algorithm, None)
+        self.assertEqual(result.versioning_status, "Enabled")
+        
+
+        key = self.random_key()
+        content = random_bytes(128 * 1024)
+
+        parts = []
+        upload_id = bucket.init_multipart_upload(key).upload_id
+
+        headers = {'Content-Md5': oss2.utils.content_md5(content)}
+
+        result = bucket.upload_part(key, upload_id, 1, content, headers=headers)
+        parts.append(oss2.models.PartInfo(1, result.etag, size=len(content), part_crc=result.crc))
+        self.assertTrue(result.crc is not None)
+
+        complete_result = bucket.complete_multipart_upload(key, upload_id, parts)
+
+        object_crc = calc_obj_crc_from_parts(parts)
+        self.assertTrue(complete_result.crc is not None)
+        self.assertEqual(object_crc, result.crc)
+        self.assertTrue(complete_result.versionid is not None)
+
+        bucket.delete_object(key, params={'versionId': complete_result.versionid})
+
+        try:
+            bucket.delete_bucket()
+        except:
+            self.assertFalse(True, "should not get a exception")
+
+    def test_upload_part_copy_with_versioning(self):
+
+        from oss2.models import BucketVersioningConfig
+        from oss2.models import BatchDeleteObjectVersion
+        from oss2.models import BatchDeleteObjectVersionList
+
+        auth = oss2.Auth(OSS_ID, OSS_SECRET)
+        bucket_name = OSS_BUCKET + "-test-upload-part-copy-with-versioning"
+        bucket = oss2.Bucket(auth, OSS_ENDPOINT, bucket_name)
+
+        bucket.create_bucket(oss2.BUCKET_ACL_PRIVATE)
+
+        wait_meta_sync()
+
+        config = BucketVersioningConfig()
+        config.status = 'Enabled'
+
+        result = bucket.put_bucket_versioning(config)
+
+        wait_meta_sync()
+
+        result = bucket.get_bucket_info()
+
+        self.assertEqual(int(result.status)/100, 2)
+        self.assertEqual(result.bucket_encryption_rule.sse_algorithm, None)
+        self.assertEqual(result.versioning_status, "Enabled")
+
+        src_object = self.random_key()
+        dst_object = self.random_key()
+
+        content = random_bytes(200 * 1024)
+        content2 = random_bytes(200 * 1024)
+
+        # 上传源文件 version1
+        put_result1 = bucket.put_object(src_object, content)
+        self.assertTrue(put_result1.versionid is not None)
+        versionid1 = put_result1.versionid
+
+        # 上传源文件 version2
+        put_result2 = bucket.put_object(src_object, content2)
+        self.assertTrue(put_result2.versionid is not None)
+        versionid2 = put_result2.versionid
+
+        # part copy到目标文件
+        parts = []
+        upload_id = bucket.init_multipart_upload(dst_object).upload_id
+
+        result = bucket.upload_part_copy(bucket_name, src_object,
+                                              (0, 100 * 1024 - 1), dst_object, upload_id, 1)
+        parts.append(oss2.models.PartInfo(1, result.etag))
+
+        result = bucket.upload_part_copy(bucket_name, src_object,
+                        (100*1024, None), dst_object, upload_id, 2, params={'versionId': versionid1})
+
+        parts.append(oss2.models.PartInfo(2, result.etag))
+
+        complete_result = bucket.complete_multipart_upload(dst_object, upload_id, parts)
+
+        # 验证
+        content_got = bucket.get_object(dst_object).read()
+        self.assertEqual(len(content_got), len(content))
+        self.assertTrue(content_got != content)
+
+        version_list = BatchDeleteObjectVersionList()
+        version_list.append(BatchDeleteObjectVersion(key=src_object, versionid=versionid1))
+        version_list.append(BatchDeleteObjectVersion(key=src_object, versionid=versionid2))
+        version_list.append(BatchDeleteObjectVersion(key=dst_object, versionid=complete_result.versionid))
+
+        self.assertTrue(version_list.len(), 3)
+
+        result = bucket.delete_object_versions(version_list)
+
+        try:
+            bucket.delete_bucket()
+        except:
+            self.assertFalse(True, "should not get a exception")
+    
 if __name__ == '__main__':
     unittest.main()
