+++ tests/test_object_request_payment_versions.py
@@ -12,9 +12,6 @@ from oss2.models import (PAYER_REQUESTER,
 class TestObjectRequestPaymentVersions(OssTestCase):
     def setUp(self):
         OssTestCase.setUp(self)
-        self.endpoint = "http://oss-ap-south-1.aliyuncs.com"
-        bucket_name = OSS_BUCKET + "-test-request-payment-versionging"
-
         policy_text = ''
         policy_text += '{'
         policy_text += '"Version":"1",'
@@ -22,25 +19,20 @@ class TestObjectRequestPaymentVersions(OssTestCase):
         policy_text += '"Action":["oss:*"],'
         policy_text += '"Effect":"Allow",'
         policy_text += '"Principal":["{0}"],'.format(OSS_PAYER_UID)
-        policy_text += '"Resource": ["acs:oss:*:*:{0}","acs:oss:*:*:{0}/*"]'.format(bucket_name)
+        policy_text += '"Resource": ["acs:oss:*:*:{0}","acs:oss:*:*:{0}/*"]'.format(OSS_BUCKET)
         policy_text += '}]}'
-
-        auth = oss2.Auth(OSS_ID, OSS_SECRET)
-        self.owner_bucket = oss2.Bucket(auth, self.endpoint, bucket_name)
-        self.owner_bucket.create_bucket(oss2.BUCKET_ACL_PRIVATE)
-
-        self.owner_bucket.put_bucket_policy(policy_text)
+        self.bucket.put_bucket_policy(policy_text)
 
         # Enable bucket request payment
-        result = self.owner_bucket.put_bucket_request_payment(PAYER_REQUESTER)
+        result = self.bucket.put_bucket_request_payment(PAYER_REQUESTER)
         self.assertEqual(result.status, 200)
 
-        self.payer_bucket = oss2.Bucket(oss2.Auth(OSS_PAYER_ID, OSS_PAYER_SECRET), self.endpoint, bucket_name)
+        self.payer_bucket = oss2.Bucket(oss2.Auth(OSS_PAYER_ID, OSS_PAYER_SECRET), OSS_ENDPOINT, OSS_BUCKET)
 
         # Enable bucket versioning
         config = BucketVersioningConfig()
         config.status = oss2.BUCKET_VERSIONING_ENABLE
-        self.owner_bucket.put_bucket_versioning(config)
+        self.bucket.put_bucket_versioning(config)
         self.assertEqual(result.status, 200)
 
     def test_delete_object_versions(self):
@@ -48,10 +40,10 @@ class TestObjectRequestPaymentVersions(OssTestCase):
         content1 = 'test-content-1'
         content2 = 'test-content-2'
 
-        result = self.owner_bucket.put_object(key, content1)
+        result = self.bucket.put_object(key, content1)
         versionid1 = result.versionid
 
-        result = self.owner_bucket.put_object(key, content2)
+        result = self.bucket.put_object(key, content2)
         versionid2 = result.versionid
 
         version_list = BatchDeleteObjectVersionList()
